Added loading screen on superuser/contests/create and superuser/contests/:contestCode/edit routes
