+++ src/Components/superuser/createContest/CreateContest.jsx
@@ -1,5 +1,4 @@
 import React, { useState } from 'react';
-import PropTypes from 'prop-types';
 import { useHistory } from 'react-router-dom';
 import { useApolloClient } from '@apollo/react-hooks';
 import { Headline4, Body2 } from '@material/react-typography';
@@ -8,11 +7,7 @@ import ContestDetails from './ContestDetails';
 import MessageCard from '../../common/MessageCard/index';
 import { CREATE_CONTEST } from '../../../graphql/mutations';
 
-const CreateContest = ({ users }) => {
-  const mappedUsers = users.map((admin) => ({
-    label: `${admin.name} (${admin.username})`,
-    value: admin._id,
-  }));
+const CreateContest = () => {
   const intialFormDetails = {
     code: '',
     type: String(/RATED/).substring(1).slice(0, -1),
@@ -32,7 +27,6 @@ const CreateContest = ({ users }) => {
   const updateDB = async () => {
     setMessageType('loading');
     setMessage('Creating Contest, Please wait');
-    // console.log(String(formDetails.type).substring(1).slice(0,-1));
     const selectedAdmins = formDetails.admins.map((admin) => admin.value);
 
     const { data, error } = await client.mutate({
@@ -69,21 +63,13 @@ const CreateContest = ({ users }) => {
     <div className="mw7 center pa2">
       <Headline4 className="purple mb1 mt3">Create Contest</Headline4>
       <Body2 className="mt0 mid-gray mb4">Create a Single Round Match or Long Queue Contest</Body2>
-      <ContestDetails
-        formDetails={formDetails}
-        setFormDetails={setFormDetails}
-        adminOptions={mappedUsers}
-      />
+      <ContestDetails formDetails={formDetails} setFormDetails={setFormDetails} />
       <MessageCard messageType={messageType} message={message} setMessageType={setMessageType} />
       <Button className="ma1 ml0 mt3" raised onClick={updateDB}>
         Create Contest
       </Button>
     </div>
   );
-};
-
-CreateContest.propTypes = {
-  users: PropTypes.array.isRequired,
 };
 
 export default CreateContest;
